Section17Optimization

+++ API/SignalR/MessageHub.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.SignalR;
 namespace API.SignalR
 {
     public class MessageHub(IMessageRepository messageRepository,
-     IUserRepository userRepository, IMapper mapper) : Hub
+     IUserRepository userRepository, IMapper mapper, IHubContext<PresenceHub> presenceHub ) : Hub
     {
         public override async Task OnConnectedAsync()
         {
@@ -19,18 +19,22 @@ namespace API.SignalR
                 throw new Exception("Cannot join group");
             var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
             await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-            await AddToGroup(groupName);
+            var group = await AddToGroup(groupName);
+
+            await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
 
             var messages = await messageRepository
                 .GetMessageThread(Context.User.GetUsername(), otherUser!);
                 
 
-            await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+            await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
+
         }
 
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
-            await RemoveFromMessageGroup();
+            var group = await RemoveFromMessageGroup();
+            await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
             await base.OnDisconnectedAsync(exception);
         }
 
@@ -64,6 +68,16 @@ namespace API.SignalR
             {
                 message.DateRead = DateTime.UtcNow;
             }
+            else
+            {
+                var connections = await PresenceTracker.GetConnectionsForUser(recipient.UserName);
+                if(connections != null && connections?.Count != null)
+                {
+                    await presenceHub.Clients.Clients(connections).SendAsync("NewMessageReceived",
+                    new {username = sender.UserName, knownAs = sender.KnownAs, message = mapper.Map<MessageDto>(message)});
+                }
+                
+            }
               
 
             messageRepository.AddMessage(message);
@@ -76,7 +90,7 @@ namespace API.SignalR
             
         }
 
-        private async Task<bool> AddToGroup(string groupName)
+        private async Task<Group> AddToGroup(string groupName)
         {
            var username = Context.User?.GetUsername() ?? throw new Exception("User not found");
            var group = await messageRepository.GetMessageGroup(groupName);
@@ -88,17 +102,21 @@ namespace API.SignalR
               messageRepository.AddGroup(group);
            }
               group.Connections.Add(connection);
-              return await messageRepository.SaveAllAsync();
+              if(await messageRepository.SaveAllAsync()) return group;
+                throw new HubException("Failed to add to group");
         } 
 
-        private async Task RemoveFromMessageGroup()
+        private async Task<Group> RemoveFromMessageGroup()
         {
-            var connection = await messageRepository.GetConnection(Context.ConnectionId);
-            if (connection != null)
+            var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+            var connection = group?.Connections.FirstOrDefault(x => x.ConnectionId == Context.ConnectionId);
+            if (connection != null && group !=null)
             {
                 messageRepository.RemoveConnection(connection);
-                await messageRepository.SaveAllAsync();
+                if(await messageRepository.SaveAllAsync()) return group;
             }
+
+            throw new HubException("Failed to remove from group");
         }
 
         private string GetGroupName(string caller, string? other)
